Add files via upload
Final copies of all the files. Includes the binary tree files in their final form, the makefile, as well as a script with 5 different test functions: 1 for output, 1 for missing argument, 1 for non-existent file, 1 for working initialization and 1 for testing search and insert. Once you clone all of the files you should be able to just run the script in the terminal using ./script2 and it should print out all passing tests. If anything goes wrong while your checking if it works just let me know!